zuul: drop secret from stable branch
Zuul does not allow secrets which are not identical acros branches,
therefore if we want to rotate our current secrets, we have to drop
all the old ones before the new one can be added.

The secret will continue to be read from the master branch, even for
the stable branch jobs so it can continue to live there.

Change-Id: I3af760c51029175be042fb9cff3cb37bfbbeefe5
Needed-By: I4f913bc06a68cf0f3d61d836ab8a7acae54d50f7