Allow for cluster-autoscaler deployment roll-out
hostNetwork: True ties the pod to the host ip:port.

When a CA update roll-out is triggered, the new pod cannot bind to the
port because it is already in use by the pod to be replaced, making the
new pod fail, and in consequence the update.

This patch removes this restriction

Story: 2006161
Task: 35670

Change-Id: I039b2ed91c9623883635bf9e72d85135f1bdae89
Signed-off-by: Diogo Guerra <dy090.guerra@gmail.com>