Fix addon tag/version parsing
When doing rolling upgrade, the addon tag/version will be parsed
to make sure there is no downgrade. But if there is tag/version
is not well-formated, it can't be parsed by pbr.version.SemanticVersion.
This patch adds a catch to avoid throwing error and just skip
this case.

Task: 36186
Story: 2002210

Change-Id: I846cd0fd40a6607c36fff8992d98d8a55c49b3fa