Remove unneeded Zuul branch matcher
We have implicit branch matchers, so there's no need to add a check for
not-ocata, this job is only run for the branch it's on - like master
now.

Remove it to not confuse Zuul when multiple branches matches and the job
is different.

Also fix heat-container-agent to make linking of /usr/bin/python to python3
conditional on whether it already exists.

Change-Id: I1fec9d19cd14cf2fe2473b7610870e8d669021b9