zuul: drop secret from stable branch
Zuul does not allow secrets which are not identical acros branches,
therefore if we want to rotate our current secrets, we have to drop
all the old ones before the new one can be added.

The secret will continue to be read from the master branch, even for
the stable branch jobs so it can continue to live there.

Change-Id: I2b64bf9819e3bfa81d9d0220a6ba7c25e9798d53
Signed-off-by: Spyros Trigazis <spyridon.trigazis@cern.ch>