bug: Only query Cinder API if volume size > 0
Only query Cinder API for volume type if volume size is greater than 0.
This is to make sure that Magnum does not have a hard dependency on
Cinder and works without the Cinder API enlisted as an endpoint.

This regression was introduced in https://review.opendev.org/#/c/621734/
where boot from volume support was introduced. See
https://storyboard.openstack.org/#!/story/2005386#comment-147555 for
more details of the error.

Story: 2005386
Task: 37565

Change-Id: I1bdec0c11a18bea9143fb4792a0a2e4ba5dc1fcf