Change k8s-keystone-auth docker repo
Reverts the change in https://review.opendev.org/#/c/643225/22 (line 9).
This allows users to deploy clusters with all released versions of
cloud-provider-openstack. The openstackmagnum repo only has v1.14.0

Also updates the docs with the default value for the Train/Ussuri release.

Updates the default tag to v1.15.0 for Ussuri.

Story: 2006967
Change-Id: Icf6f9f3c02d6b211c4dc3c1d654c8d8573443e0f