[hca] Restore deploy_{stdout,stderr,status_code}
In I5504c00efce89105d403722d583bb75f7bdea714, we removed
deploy_stderr from the output and instead piped everything into
deploy_stdout. Turns out that this is not backward compatible and
removes it from the ouptut of:

    openstack software deployment output show <output_id> --long --all

This PS uses threading to write stdout and stderr live to a file and
correctly return deploy_status_code instead of None.

Story: 2007264
Task: 38983

Change-Id: I174e80c6982317f52150a4b255f3d1c592d9caaf
Signed-off-by: Bharat Kunwar <brtknr@bath.edu>