Use cluster name for fixed_network instead of private
At present, when a fixed_network is not specified, it is given the name
"private" by default. When multiple clusters are created, we end up in a
situation where we end up with multiple networks all with the same name.
This PS intends to make it easier to see where the resources belong to
by using the cluster name.

Story: 2007460
Task: 39139

Change-Id: I7f8028b716f9a9eced17d85ca2e46e2b1e34875f