[k8s] Fix docker storage of Fedora CoreOS
In commit I1a75f1bf12747508a3497293650d3cc668202de6 the worker node
is missed to add the docker storage support. And the current systemd
unit is not really working. So this patch fixes it by removing the
hardcode for /dev/vdb and using xfs instead of ext4 (the same way
for Fedora Atomic) to make it simpler and solid.

Task: 39331
Story: 2005201

Change-Id: I4c465664eb19f1992df95750dd7b2d99688c6cae