[k8s] Fix no IP address in api_address
This is a corner case that when floating_ip_enabled=False,
master_lb_enabled=True,master_lb_floating_ip_enabled=False in
cluster template, but setting floating_ip_enabled=True when
creating the cluster. The current logic is not correct which
resulted in missing IP address in the api_address of cluster.

Task: 39519
Story: 2007586

Change-Id: I5e2ca270c4f4e2c48d067cd5b8f6609c037cb6e5