Deprecation note for devicemapper and overlay
The devicemapper and overlay storage driver are deprecated in
favor of overlay2 in docker, and will be removed in a future
release from docker. Users of the devicemapper and overlay
storage driver are recommended to migrate to a different
storage driver, such as overlay2. overlay2 will be set as
the default storage driver from Victoria cycle.

Task: 39526
Story: 2005201

Change-Id: Idfd784dc3f1bfbe6bce5671bc5a1ef69d82636b1