Remove cloud-config from k8s worker node
Now Magnum is not deploying any service or workload on k8s
worker nodes which need to get credentials from local to talk
to Magnum control plane. So the cloud-config file should be
removed from worker nodes to reduce the attach surface from
a security point of view.

Task: 40791
Story: 2008090

Change-Id: I72e418491cbd19291527bbe4b504d599c740fea9