ClusterAPI: implement update_cluster_status
This patch means the ClusterAPI driver can now successfully create
and delete clusters.

You can see this being tested in Zuul in these
patches which change what we test to include cluster create:
https://review.opendev.org/c/openstack/magnum-tempest-plugin/+/884366
https://review.opendev.org/c/openstack/magnum-tempest-plugin/+/872759

We add checks to see if the helm chart application has actually
worked, by looking at the capi resoruces in the capi management cluster,
and then update the DB state so the API tells the users the cluster is ready.

We can now create and delete clusters using the new driver!
To be able to test this within devstack we have added a CAPI k8s cluster
on the controller node that we will use to create all the magnum clusters.

To aid developer and reviewer testing we have added a script to bring
up this environment within devstack:
devstack/contrib/new-devstack.sh

Future patches will fix up other APIs such as resizing the cluster
and adding additional node pools.

story: 2009780

Co-Authored-By: Matt Pryor<matt@stackhpc.com>
Co-Authored-By: Tyler Christie<tyler@stackhpc.com>
Co-Authored-By: John Garbutt<johng@stackhpc.com>

Change-Id: If583bec63a2135486e393bacb16402f6821b7ede