Fix test for barbican cached client
Test is broken after Change Ibfaea3fe9e394f6b1286d92437629d0400305968 in
barbicanclient. As this test mock keystoneclient, version discovery by
barbicanclient using the MagicMock keystone session now fails.

As this test just wants to make sure we are using a cached
barbicanclient, it should be safe to mock barbicanclient and check that
the client is only initialised once.

Change-Id: I85e70ea637af9c0b467aa32baf107d7fbed914c0