Update gate jobs as per the 2025.1 cycle testing runtime
As per 2025.1 testing runtime[1], we need to test on Ubuntu
Noble (which will be taken care by depends-on tempest and devstack
patches to move base jobs to Noble) and at least single job to run on
Ubuntu Jammy (for smooth upgrade from previous releases).

This commit adds a new job to run on Jammy which can be removed
in future cycle when testing runtime test next version of Ubuntu
as default.

Depends-On: https://review.opendev.org/c/openstack/magnum-tempest-plugin/+/932954

[1] https://governance.openstack.org/tc/reference/runtimes/2025.1.html

Change-Id: Iee421d4c2475e5f46bb1d270086a0c793e4c0fe4