api: Respect all service type aliases
The Service Types Authority indicates [1] that the official service type
for Magnum is 'container-infrastructure-management'. While all clients
are using the alias 'container-infra', and will have to continue doing
so for a long time to work with older Magnum deployments, we should
support the official service type along with any other aliases.

[1] https://specs.openstack.org/openstack/service-types-authority/

Change-Id: I7dee4471702da513a0e6f231504b731aea94395a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>