api: Use application_url, not host_url
Per the webob docs [1], 'host_url' returns the URL of the host rather
than the URL of the application (which is the URL of the host plus the
'SCRIPT_NAME' [2] portion of the path). This means if you deploy Magnum
under a path, like DevStack now does, you are missing part of the path.
Correct this by using 'application_url' instead, like Nova [3] and other
projects already do.

Note that this is compliant with the version discovery guidelines, which
indicate that absolute paths can be consumed wholesale [4].

[1] https://docs.pylonsproject.org/projects/webob/en/stable/index.html#urls
[2] https://peps.python.org/pep-3333/
[3] https://opendev.org/openstack/nova/src/commit/64ca204c9cf497b0dcfff2d3a24b0dd795a57d1d/nova/api/openstack/common.py#L389-L414
[4] https://specs.openstack.org/openstack/api-sig/guidelines/consuming-catalog/version-discovery.html#expanding-endpoints

Change-Id: I42da06cd1b04e05ac2c4224c06992ba630854fac
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes-bug: #2083168